Minor fixes (#367)
This PR is made for small changes I see while working on other issues.

- Improve `.gitignore` to ignore files like `development.sqlite3-shm`
and `development.sqlite3-wal` in `spec/dummy/db/`.
- Update the version of `pry-doc` (and other related gems) in the dummy
app `Gemfile.lock` to work with Ruby `3.3.0`.
- Update the wrong Procfile name in Profile so that the message shown by
running `foreman start` prints correct Procfile names.
- Remove `local` script from `package.json`. This script is faulty and
needs to be fixed if we still want to use it. It is removed because
there is no reference to it, and we have no report on its faulty
behavior, meaning we don't use it at all.
- Prevent storing rspec results when running the tests locally (to
prevent the creation of `~` directory in the root of the project)